+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change 'need_bootstrappers' to have dinamically threshold [#135]
+- Change `BinaryID::from_nonce` to return result [#136]
 
 ## [0.6.0] - 2023-11-01
 
@@ -135,6 +136,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#117]: https://github.com/dusk-network/kadcast/issues/117
 [#123]: https://github.com/dusk-network/kadcast/issues/123
 [#135]: https://github.com/dusk-network/kadcast/issues/135
+[#136]: https://github.com/dusk-network/kadcast/issues/136
 
 <!-- Releases -->